fix(twai): Tx interrupt handler loads next frame after bus-off
This commit fixes an issue in "twai_handle_tx_buffer_frame()" where the
handler will attempt to load another message from the TX queue for
transmission even if in a bus-off state. Combined with a subsequent call to
"twai_initiate_recovery()", the following interrupt on completion of bus
recovery will result in "tx_msg_count" going negative, result in a failed
assert.

Closes https://github.com/espressif/esp-idf/issues/9697